requirements.txt
xrpl-py
python-dotenv
Flask
ecdsa
bech32
shortuuid
Flask-SQLAlchemy
Flask-Migrate
pyjwt